Builds with all warnings enabled and just a few disabled (due to issues in the frameworks)

+++ CmdParser4Cpp/src/Argument.cpp
@@ -81,7 +81,7 @@ Argument::FindArgument( const VectorOfString& allArguments, int& hitCount )
 	// Loop all our names and test against all provided arguments.
 	for( const auto& name : myNames ) {
 		for( int ix = 0; ix < static_cast<int>( allArguments.size() ); ++ix ) {
-			if( allArguments.at( ix ).compare( name ) == 0 ) {
+			if( allArguments.at( static_cast<size_t>( ix ) ).compare( name ) == 0 ) {
 				if( foundIx == -1 ) {
 					// First hit
 					foundIx = ix;
